M3: Created Portfolio Class
closes #32

Implemented the Portfolio class according to the class diagram, except the cashflow.

Additional functions will get added when the need for them arises but most of them should be implemented in the PortfolioService instead of the Portfolio itself.

For the mappings from Portfolio to Postion to work I had to create the Position class and implement two fields there. Therefore the Portfolio class is not finished yet.