M3: Safely connected to Heroku Postgres
The Database credentials are hidden now so they are only visible to heroku.
That means that for local testing we can use a local TEMPORARY DB. To do that we need to uncomment spring.profile.active=local.

To use the persistent database locally you need to input the database credentials into either local properties or application properties.

DO NOT UNDER ANY CIRCUMSTANCES COMMIT APPLICATION.PROPERTIES OR APPLICATION-LOCAL.PROPERTIES WITH THE REAL DATABASE CREDENTIALS.